+++ .github/workflows/rust.yml
@@ -6,8 +6,8 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v1
-    - name: Build
-      run: cargo build --verbose
+      - uses: actions/checkout@v1
+      - name: Build
+        run: cargo build --verbose
     # - name: Run tests
-    #   run: cargo test --verbose
+    #   run: cargo test --no-fail-fast --verbose
