+++ tests/misc.rs
@@ -1,0 +1,16 @@
+/*
+#[test]
+fn test_thaw_wrong_object() {
+    assert!(false);
+}
+
+#[test]
+fn test_missing_headers() {
+    assert!(false);
+}
+
+#[test]
+fn test_github_response_with_small_clock_skew() {
+    assert!(false);
+}
+*/
